Day 10, complete.
I understand that it has to do with the patterns of differences.

I suspect it is a discrete recurrence of some sort.

I recognized the pattern (ish) last night.

I don't like how I solved it, which was by hard-coding the relationships
into the determination of how many previous values to sum...

Perhaps I'll look at other solutions at some point, and either I'll have
found a reasonable way to do it, or I'll have found the hard way.