Conway's GoL
I've got it working for the small map, but my implementation is too slow
for the large map.

I left it running while going to the grocery story, and it didn't
finish. I flipped it over to a vector representation, but it's still too
slow, even without copying/creating new maps for every iteration. (I'm
essentially double-buffering, flopping back-and-forth between maps.)

However, it's still too slow?