Exploring bitwise representations...
The board could be represented as a 100x100 bitstring. All operations on
the board would then be maps that were OR/ANDed against it.

This... should be faster?

For now, I'm going to run it and see what happens. I have a feeling
something is looping, or I have something that is correct for the small
map, but I'm discovering an error in the large map?